+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Use bg-secondary for selected model card
 - Use ResponsivePopover for model config, fix card width
 
 ### Fixed